refactor: replace bundler from react-script to Vite
