+++ .gitignore
@@ -13,12 +13,12 @@
 
 # misc
 .DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
+.env
 .idea
 
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# Generated subgraph (GraphQL) client
+src/subgraph/client
