+++ index.html
@@ -42,11 +42,11 @@
     -->
     <link rel="manifest" href="https://venus.io/manifest.json" />
     <!--
-      Notice the use of https://venus.io in the tags above.
+      Notice the use of  in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
       Only files inside the `public` folder can be referenced from the HTML.
 
-      Unlike "/favicon.ico" or "favicon.ico", "https://venus.io/favicon.ico" will
+      Unlike "/favicon.ico" or "favicon.ico", "/favicon.ico" will
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
@@ -67,7 +67,7 @@
       content="Venus is a decentralized finance (DeFi) algorithmic money market protocol on BNB Chain."
     />
     <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:image" content="https://venus.io/share.png" />
+    <meta name="twitter:image:src" content="https://venus.io/share.png" />
     <meta name="theme-color" content="#3A78FF" />
 
     <meta property="og:title" content="Venus Protocol" />
@@ -94,5 +94,7 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
+
+    <script type="module" src="/src/index.tsx"></script>
   </body>
 </html>
