+++ package.json
@@ -3,34 +3,29 @@
   "homepage": "https://venus.io/",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
+    "@tanstack/react-query": "^5.40.0",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^16.7.13",
-    "@types/react": "^17.0.40",
-    "@types/react-dom": "^17.0.13",
     "axios": "^0.26.1",
     "classnames": "^2.3.1",
     "normalize.css": "^8.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-particles": "2.9.3",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.0",
     "react-scrolllock": "^5.0.1",
     "tsparticles": "2.9.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-app-rewired start",
-    "predeploy": "npm run build",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "lint:fix": "yarn lint --fix && yarn pretty --write",
-    "lint": "eslint .",
+    "lint": "yarn run eslint .",
     "pretty": "yarn prettier \"src/**/*.{js,jsx,ts,tsx,html,json,md}\" public/**/*.html --check"
   },
   "browserslist": {
@@ -46,12 +41,22 @@
     ]
   },
   "devDependencies": {
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.12.12",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "^7.11.0",
+    "@typescript-eslint/parser": "^7.11.0",
+    "@vitejs/plugin-react": "^4.3.1",
     "autoprefixer": "^9.0.0",
+    "eslint": "^8",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^16.1.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-config-airbnb-typescript": "^18.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.34.2",
     "husky": "^7.0.4",
     "lint-staged": "^12.3.5",
     "postcss": "^8.4.8",
@@ -64,7 +69,10 @@
     "react-app-rewire-postcss": "^3.0.2",
     "react-app-rewired": "^2.2.1",
     "react-snap": "^1.23.0",
-    "typescript": "^4.6.2"
+    "typescript": "^5.4.5",
+    "vite": "^5.3.1",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-tsconfig-paths": "^4.3.2"
   },
   "resolutions": {
     "postcss": "^8.4.8"
