+++ src/components/App.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import MainContent from './MainContent/MainContent';
 import Footer from './Footer/Footer';
 import s from './App.module.css';
@@ -13,6 +14,8 @@ function Main() {
   );
 }
 
+const queryClient = new QueryClient();
+
 function App() {
   useEffect(() => {
     if (window.location.pathname.startsWith('/discord')) {
@@ -21,13 +24,15 @@ function App() {
   }, [window.location.pathname]);
 
   return (
-    <BrowserRouter>
-      <Routes>
-        {['/', '/discord'].map(path => (
-          <Route path={path} element={<Main />} key={path} />
-        ))}
-      </Routes>
-    </BrowserRouter>
+    <QueryClientProvider client={queryClient}>
+      <BrowserRouter>
+        <Routes>
+          {['/', '/discord'].map(path => (
+            <Route path={path} element={<Main />} key={path} />
+          ))}
+        </Routes>
+      </BrowserRouter>
+    </QueryClientProvider>
   );
 }
