+++ src/components/Footer/Footer.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import cn from 'classnames';
 import Container from '../Container/Container';
 import NavigationLinks from '../NavigationLinks/NavigationLinks';
-import { ReactComponent as Logo } from '../Header/assets/logo.svg';
-import { ReactComponent as LogoMobile } from './assets/logoSmall.svg';
+import Logo from '../Header/assets/logo.svg?react';
+import LogoMobile from './assets/logoSmall.svg?react';
 import SocialLinks from './SocialLinks';
 import s from './Footer.module.css';
 import Link from '../Link/Link';
