+++ vite.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import viteTsconfigPaths from 'vite-tsconfig-paths';
+import svgr from 'vite-plugin-svgr';
+
+export default defineConfig({
+  plugins: [react(), viteTsconfigPaths(), svgr()],
+  build: {
+    sourcemap: true,
+    outDir: 'build',
+  },
+});
